feat: refactoring to_dict to modular
