+++ app/v1/modules/kubernetes_services/__init__.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+
+
+from .script_to_dict import ScriptToDictParserService
+
+
+__all__ = [
+    "ScriptToDictParserService",
+]
